Add compatibility to UniFi Controller 5.6
The controller adds a unifi user and group which uses different UID/GID.
So we install the user/group before UniFi controller so our UID/GID gets
preserved.

Needs to correct a warning on startup with "operation not permitted".
This warnign can safely be ignored as permissions are correct. But it is
obviously not nice.